pushing from 100 days of python code exercise
This is from my project that is assuming, if you live until 90, how many days, weeks, and months from a given age, that you would have left until you reach the age of 90. This helped give me some perspective on the topic of time, and what it truly means to know the clock is ticking. 

This article gives more insight: https://waitbutwhy.com/2014/05/life-weeks.html